Re-visiting this good old project
This was my first attempt to create a text-based planting game 4 years ago when I was still learning Python. This was written by a different developer. Seeing this old project make me reminisce the moments I've been through. Man, time flies so freakin' fast when you are enjoying it :smile: