Created mechanics.py for game, player, inventory, etc.
So far I managed to implement the whole codebase in a class-based approach. Here are the list of mechanics of leaf.

`GameMechanics` - game related tasks like, running, showing help and quiting the application.
`PlayerMechanics` - user related actions like 'till', 'plant', 'harvest', etc.
`InventoryMechanics` - items that can be collected by the user and use in the game.