Implemented the 'harvest' command and simple gameplay
The essential commands are now implemented: till > plant > check > harvest

Other features and mechanics will be added around these commands, see added TODOs. Tests will be implemented on the coming coding sessions.

So far, the game is now working as expected but there are still a lot more logic errors to fix and make the game more intuitive to the player.