Implemented and added new inventory commands but with logic error
Commands residing on a same mechanics class seems to work as intended but when another game command tries to access a variable/method from another class, the result is not what I intended.

For example:
# add item to inventory and display result
> add_item tomato
> inventory
tomato

# but plant command can't see the added new item
> plant tomato
You don't have a 'tomato' in your inventory.

I have assumptions for this logic error but I think I need to focus more testing and keep the code of execution as simple sa possible for this very simple text-based game!