Solved InventoryMechanics' last week logic error
I was re-initializing InventoryMechanics class in both the GameMechanics and PlayerMechanics init methods. That lead to resetting the values of the InventoryMechanics' inventoryDeque variable every time I call the `plant` commands.

Simply removing this line from the PlayerMechanics init method and transferring it to the GameMechanics init method solves the problem:

self.playerInventoryMechanics = InventoryMechanics()

We can now move on to some other tasks, *whew*