Prevented user from planting by adding isTilled flag
Player needs to use first the 'till' command before planting. The isTilled will be reset to false for every complated plant command.

This is bad implementation because of the nested ifs but it does the work of preventing the user from planting. I will definitely refactor in the next coding session.